+++ golang/btcspv/utils.go
@@ -1,0 +1,23 @@
+package btcspv
+
+import (
+	"encoding/hex"
+)
+
+func strip0xPrefix(s string) string {
+	if len(s) < 2 {
+		return s
+	}
+	if s[0:2] == "0x" {
+		return s[2:]
+	}
+	return s
+}
+
+func DecodeIfHex(s string) []byte {
+	res, err := hex.DecodeString(strip0xPrefix(s))
+	if err != nil {
+		return []byte(s)
+	}
+	return res
+}
